+++ src/libaktualizr/storage/sqlstorage.cc
@@ -1416,7 +1416,41 @@ std::unique_ptr<StorageTargetRHandle> SQLStorage::openTargetFile(const Uptane::T
   return std_::make_unique<SQLTargetRHandle>(*this, target);
 }
 
-void SQLStorage::removeTargetFile(const std::string& filename) {
+std::vector<Uptane::Target> SQLStorage::getTargetFiles() {
+  SQLite3Guard db = dbConnection();
+
+  auto statement = db.prepareStatement<>("SELECT targetname, real_size, sha256, sha512 FROM target_images;");
+
+  std::vector<Uptane::Target> v;
+
+  int result = statement.step();
+  while (result != SQLITE_DONE) {
+    if (result != SQLITE_ROW) {
+      LOG_ERROR << "Statement step failure: " << db.errmsg();
+      throw std::runtime_error("Error getting target files");
+    }
+
+    auto tname = statement.get_result_col_str(0).value();
+    auto tsize = statement.get_result_col_int(1);
+    auto sha256 = statement.get_result_col_str(2).value();
+    auto sha512 = statement.get_result_col_str(3).value();
+
+    std::vector<Uptane::Hash> hashes;
+    if (!sha256.empty()) {
+      hashes.emplace_back(Uptane::Hash::Type::kSha256, sha256);
+    }
+    if (!sha512.empty()) {
+      hashes.emplace_back(Uptane::Hash::Type::kSha512, sha512);
+    }
+    v.emplace_back(tname, Uptane::EcuMap{}, hashes, static_cast<size_t>(tsize));
+
+    result = statement.step();
+  }
+
+  return v;
+}
+
+void SQLStorage::removeTargetFile(const std::string& target_name) {
   SQLite3Guard db = dbConnection();
 
   if (!db.beginTransaction()) {
@@ -1425,20 +1459,27 @@ void SQLStorage::removeTargetFile(const std::string& filename) {
   }
 
   auto statement =
-      db.prepareStatement<std::string>("SELECT targetname FROM target_images WHERE targetname = ?;", filename);
+      db.prepareStatement<std::string>("SELECT filename FROM target_images WHERE targetname = ?;", target_name);
 
   if (statement.step() != SQLITE_ROW) {
     LOG_ERROR << "Statement step failure: " << db.errmsg();
     throw std::runtime_error("Could not find target file");
   }
 
-  statement = db.prepareStatement<std::string>("DELETE FROM target_images WHERE targetname=?;", filename);
+  std::string filename = statement.get_result_col_str(0).value();
+
+  statement = db.prepareStatement<std::string>("DELETE FROM target_images WHERE targetname=?;", target_name);
 
   if (statement.step() != SQLITE_DONE) {
     LOG_ERROR << "Statement step failure: " << db.errmsg();
     throw std::runtime_error("Could not remove target file");
   }
-  boost::filesystem::remove(images_path_ / filename);
+  try {
+    boost::filesystem::remove(images_path_ / filename);
+  } catch (std::exception& e) {
+    LOG_ERROR << "Could not remove target file";
+    throw;
+  }
 
   db.commitTransaction();
 }